Add the encryption support.
The key is located at 0x2000 in program memory. Find some way to write
it there. A possible way is:
* Erase the flash
* Write the bootloader
* Encrypt the key using a key of 0xFFFFFF...
* This will match the "empty key" in the flash, which should make the
bootloader happy.

Key size is set in the linker files. Make a choice and stick to it. If
that isn't flexible enough, we can write the actual size before the key
in flash, and read it from there.